Fix icon being the wrong colour on macos sometimes
